Test/pyproject.toml
[project]
name = "django_tharun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "dj-rest-auth>=5.0.2",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "requests>=2.32.5",
]


[tool.pyright]
include=["project","project/registation"]
pythonVersion="3.13"
venvPath ="."
venv=".venv"
